# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Renhead API",
        "description": "Class Controller",
        "version": "1.0.0"
    },
    "paths": {
...
        "/payment-approvals": {
            "get": {
                "tags": [
                    "Payment Approvals"
                ],
                "summary": "Get",
                "operationId": "97e6a3ec4cb36a31f2403bc242ad787d",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment Approvals"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Payment Approvals"
                ],
                "operationId": "Update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "payment_id": {
                                        "description": "Payment ID",
                                        "type": "integer"
                                    },
                                    "payment_type": {
                                        "description": "Payment Type",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "approver",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment Approvals"
                ],
                "operationId": "Store",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "payment_id": {
                                        "description": "Payment ID",
                                        "type": "integer"
                                    },
                                    "payment_type": {
                                        "description": "Payment Type",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "approver",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/payment-approvals/{id}": {
            "get": {
                "tags": [
                    "Payment Approvals"
                ],
                "summary": "Get",
                "operationId": "0809a6f5542bd33a90ed1adaf834733c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment Approvals"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payment Approvals"
                ],
                "summary": "Delete",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment Approval ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/payment/approval": {
            "post": {
                "tags": [
                    "Payment Approvals"
                ],
                "operationId": "Insert",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "payment_id": {
                                        "description": "Payment ID",
                                        "type": "integer"
                                    },
                                    "payment_type": {
                                        "description": "Payment Type",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "approver",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
}